documentation rework (#81)
We do not need manually copy the folders from the image to the mounted volume, because this is done by the entrypoint.sh script now.

Since we have no other solution for libpcap and zwave support we need to run the container as root. This is done by overriding the command in the Dockerfile. Setting the -u parameter is useless, because we use gosu.