Install ttf-dejavu for some basic fonts (needed for charts)
I've used the alpine image instead of the debian and all charts did
not work anymore. After finding:
https://github.com/docker-library/openjdk/issues/73
and installing the ttf.dejavu package manually, the charts worked again.

This commit adds this package to the alpine base packages which needs to
be installed to get a working OH2 environment.