correct timezone for containers (#185)
* correct timezone for containers

I spent some time last night wondering why my rules execute in UTC time instead of local time. Then I found the solution, and it would be nice if you could add the info on the README, as it already instructs how to set system time. Which is not enough to set it for Java. So perhaps this line could be addded:

        -e "EXTRA_JAVA_OPTS=-Duser.timezone=Europe/YourTZCity" \

Which then sets the Java to run in correct timezone, and rules work in correct times along with properly timed logs.

* YourTZCity to Berlin to allow copy paste trying

As agreed in PR comments.