Provide example playbook for Ansible in README (#186)
* Provide example playbook for Ansible in README

Ansible is more and more used to automatise server setups. I control all my boxes with it, and thought I might as well contribute my playbook running OpenHAB container with it. It could be enhanced to role, but this serves basic openhab run.

* Adding TOC entry for ansible chapter

* Changed Helsinki -> Berlin

Changed timezone cities to be aligned with the other README documentation.